+++ src/main/java/com/example/log4u/domain/media/service/MediaService.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -13,6 +14,7 @@ import com.example.log4u.domain.media.repository.MediaRepository;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import software.amazon.awssdk.services.s3.S3Client;
 
 @Service
 @RequiredArgsConstructor
@@ -20,17 +22,30 @@ import lombok.extern.slf4j.Slf4j;
 public class MediaService {
 
 	private final MediaRepository mediaRepository;
+	private final S3Client s3Client;
+
+	@Value("${S3_BUCKET_NAME}")
+	private String bucketName;
 
 	@Transactional
 	public void saveMedia(Long diaryId, List<MediaRequestDto> mediaList) {
 		if (mediaList == null || mediaList.isEmpty()) {
 			return;
 		}
-		List<Media> media = mediaList.stream()
-			.map(mediaDto -> Media.toEntity(diaryId, mediaDto))
+
+		// 미디어 ID 목록 추출
+		List<Long> mediaIds = mediaList.stream()
+			.map(MediaRequestDto::mediaId)
 			.toList();
 
-		mediaRepository.saveAll(media);
+		List<Media> existingMedia = mediaRepository.findAllById(mediaIds);
+
+		// 다이어리와 연결
+		for (Media media : existingMedia) {
+			media.connectToDiary(diaryId);
+		}
+
+		mediaRepository.saveAll(existingMedia);
 	}
 
 	@Transactional(readOnly = true)
